Upgrade rust-bitcoin
Upgrade rust-bitcoin to 0.19.2
Upgrade rust-bitcoin-rpc to 0.8.0-rc1
Remove bech32 and bitcoin_hashes dependency, use rust-bitcoin re-export
instead.

+++ cnd/src/http_api/action.rs
@@ -132,7 +132,7 @@ impl IntoResponsePayload for bitcoin::SpendOutput {
                 address,
                 fee_per_wu,
             } => {
-                let fee_per_wu = fee_per_wu.parse::<u64>().map_err(|_| {
+                let fee_per_wu = fee_per_wu.parse::<usize>().map_err(|_| {
                     HttpApiProblem::new("Invalid query parameter.")
                         .set_status(StatusCode::BAD_REQUEST)
                         .set_detail("Query parameter fee-per-byte is not a valid unsigned integer.")
